pyproject.toml
[project]
name = "deep-learning-rag-agent"
version = "0.1.0"
description = "Minimal Deep Learning RAG Interview Prep Agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.44.0",
  "chromadb>=0.5.23",
  "langchain>=0.3.20",
  "langchain-community>=0.3.20",
  "langchain-chroma>=0.2.2",
  "langgraph>=0.3.18",
  "langchain-groq>=0.2.4",
  "langchain-ollama>=0.2.3",
  "sentence-transformers>=3.4.1",
  "pydantic>=2.10.6",
  "python-dotenv>=1.0.1",
  "pypdf>=5.4.0",
  "pytest>=8.3.5"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
